+++ tests/Geopilot.Pipeline.Test/PipelineFileListTest.cs
@@ -11,10 +11,10 @@ public class PipelineFileListTest
     [TestMethod]
     public async Task FileExtensionFilter()
     {
-        PipelineFileList pipelineFiles = new PipelineFileList(new List<IPipelineFile> { roadsExdm2ienXtfFile, helloWorldPdfFile });
-        var filteredFiles = pipelineFiles.WithExtensions(new HashSet<string> { "XTF" });
-        Assert.HasCount(1, filteredFiles.Files);
-        Assert.AreEqual("RoadsExdm2ien.xtf", filteredFiles.Files.First().OriginalFileName);
+        IPipelineFile[] pipelineFiles = new IPipelineFile[] { roadsExdm2ienXtfFile, helloWorldPdfFile };
+        var filteredFiles = pipelineFiles.WithExtensions(new HashSet<string> { "XTF" }).ToList();
+        Assert.HasCount(1, filteredFiles);
+        Assert.AreEqual("RoadsExdm2ien.xtf", filteredFiles.First().OriginalFileName);
     }
 
     [TestMethod]
@@ -22,8 +22,8 @@ public class PipelineFileListTest
     [DataRow(@"^\w{1,11}\.[^.]+$", 1)]
     public async Task NameFilter(string pattern, int expectedNumber)
     {
-        PipelineFileList pipelineFiles = new PipelineFileList(new List<IPipelineFile> { roadsExdm2ienXtfFile, helloWorldPdfFile });
-        var filteredFiles = pipelineFiles.WithMatchingName(pattern);
-        Assert.HasCount(expectedNumber, filteredFiles.Files);
+        IPipelineFile[] pipelineFiles = new IPipelineFile[] { roadsExdm2ienXtfFile, helloWorldPdfFile };
+        var filteredFiles = pipelineFiles.WithMatchingName(pattern).ToList();
+        Assert.HasCount(expectedNumber, filteredFiles);
     }
 }